Images Added
Dillon has created some preliminary images that we can use!
I have added strings containing the location of the image.
The group assigned to integrating the images will need to write the
paintcomponent function to paint these images using the given string.
The image is located in the package/folder "images"