ValidationBox improvements (sort of)
The validation box now has the ability to turn red or green based on the
number entered. However, there is a bug that is consistintly making the
box turn red. I couldn't quite figure out how to fix, so if anybody
finds something wrong in the code, please tell me.